Tests/LibWeb/Text/expected/Messaging/Messaging-transfer.txt

Buffer length after transfer: 0
Port2 received undefined of data
Uncaught Error In Test: Uncaught exception: [object DOMException]
